requirements-dev.txt
-r requirements.txt

# Development tools
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
